docker: fix default network name in docker-compose.yml
If omitted, the project name is automatically deduced from the
parent directory name. This was working fine for container names
because we were explicitly defining their names, but the network
name was left unspecified. osrdyne default config depends on the
network name being "osrd_default", but that wasn't the case when
the parent directory wasn't named "osrd".

Explicitly set the project name to fix this.

Signed-off-by: Simon Ser <contact@emersion.fr>