front: display stdcm error message even when pathfinding is not called
Pathfinding is prevented when some validity checks don't pass, as for
example when the user selects the same origin and destination. The error
warning was only updated when pathfinging occurs, causing the user
warnings not to be displayed when we could detect before launching the
pathfinding that its request would be invalid.

Update the warnings independently on whether the pathfinding actually
happened so that invalid checks preventing the pathfinding also generate
their user warnings.

Signed-off-by: Loup Federico <16464925+Sh099078@users.noreply.github.com>