front: remove useSetupItinerary hook
UseSetupItinerary was used to compute the pathfinding of the train the user wants to edit. However, it was quite redundant
with the function launchPathfinding and the cohabitation of the 2 hooks created multiple bugs (power restrictions are reset
when editing a train, the invalid path items are not highligthed when editing a train etc).

From now on, the pathfinding store will only be populated from the launchPathfinding callback, even when opening the edition panel.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>