front: fix cached power restrictions on GEV
The power restrictions were computed only when the path of the selected train changed, and not when the selected train himself changed.
Thus, if 2 trains had exactly the same path, then the power restrictions were not re-computed when switching from one to another.

Now, the power restrictions are computed each time the selected train or the pathfinding changes.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>