core: fix unrecoverable freeze on exception during rabbit's DeliverCallback
Hard fix (kind of) to kill the process (and let orchestrator restart) if an
exception goes all the way up to the DeliverCallback.
For example when not able to reach editoast for infra reload.
This will release unacked messages and move them back to ready (instead
of keeping them unacked until the worker exits).

Signed-off-by: Pierre-Etienne Bougué <bougue.pe@proton.me>