front: don't use front/tests/ to write PDF files
front/tests/ is not a directory meant to be used for test results.
It should only contain source files.

Playwright sets up a specific directory for test outputs, so we
can just use that instead. As a bonus, the PDF is now included in
CI artifacts and can be inspected manually if needed.

Fixes permission errors when running tests in a container for me.

Signed-off-by: Simon Ser <contact@emersion.fr>