front: clarifies conf selectors uses
This commit aims at clarifying the way conf selectors are used now in
the applications. It makes application-specific selectors export
selectors with clear names (so that it's easy to now in which slice they
read the data), and clears STDCM and Operational Studies codebases from
any useOsrdConfSelectors calls.

It also upgrades buildCommonConfSelectors and buildInfraStateSelectors
to accept names, to return collections with better names. It is not used
yet, because since many common components still use the context
selectors, these selectors must keep common names.

Details:
- Upgrades buildCommonConfSelectors and buildInfraStateSelectors to
  give better names to their output
- Updates operationalStudiesConf/selectors to export the common
  selectors, with clearer names
- Updates stdcm/selectors similarly
- Removes all calls to useOsrdConfSelectors from Operational Studies and
  STDCM codebases

Signed-off-by: Alexis Jacomy <alexis.jacomy@gmail.com>