front: drop timestampToHHMMSS()
This function is buggy: it formats time in UTC instead of using the
local timezone.

Instead, use dateToHHMMSS(), which is also more type-safe since it
takes a Date as input instead of a number of seconds. This requires
getting rid of convertDepartureTimeIntoSec() usage as well, which
is great because that function performs some getTimezoneOffset()
shenanigans (ends up with the wrong offset depending on daylight
savings). We still need to stick to numbers instead of Date objects
because we interpolate values.

Signed-off-by: Simon Ser <contact@emersion.fr>